CS: Move data schema & generation to dapper
Move data schema & generation related code under the dapper package
where it's supposed to reside.
