+++ packages/cyberstorm/src/components/CommunityCard/CommunityCard.tsx
@@ -3,7 +3,7 @@ import { MetaItem } from "../MetaItem/MetaItem";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faBoxOpen, faDownload } from "@fortawesome/free-solid-svg-icons";
 import { formatInteger } from "../../utils/utils";
-import { CommunityPreview } from "../../schema";
+import { CommunityPreview } from "@thunderstore/dapper/src/schema";
 import { CommunityLink } from "../Links/Links";
 
 export interface GameIconProps {
