+++ packages/cyberstorm/src/components/Layout/Settings/SettingsLayout.tsx
@@ -6,9 +6,9 @@ import { SettingsLink } from "../../Links/Links";
 import { Tabs } from "../../Tabs/Tabs";
 import { Connections } from "./Connections/Connections";
 import { Account } from "./Account/Account";
-import { getUserSettingsDummyData } from "../../../dummyData";
+import { getUserSettingsDummyData } from "@thunderstore/dapper/src/implementations/dummy/generate";
 import { BaseLayout } from "../BaseLayout/BaseLayout";
-import { UserSettings } from "../../../schema";
+import { UserSettings } from "@thunderstore/dapper/src/schema";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faCircleNodes, faCog } from "@fortawesome/pro-regular-svg-icons";
 import { PageHeader } from "../BaseLayout/PageHeader/PageHeader";
