+++ packages/cyberstorm/src/components/Layout/Teams/TeamSettings/TeamServiceAccounts/TeamServiceAccounts.tsx
@@ -3,8 +3,8 @@ import { SettingItem } from "../../../../SettingItem/SettingItem";
 import { Button } from "../../../../Button/Button";
 import { ServiceAccountList } from "./ServiceAccountList/ServiceAccountList";
 import { Dialog } from "../../../../Dialog/Dialog";
-import { ServiceAccount, TeamSettings } from "../../../../../schema";
-import { getServiceAccountDummyData } from "../../../../../dummyData";
+import { ServiceAccount, TeamSettings } from "@thunderstore/dapper/src/schema";
+import { getServiceAccountDummyData } from "@thunderstore/dapper/src/implementations/dummy/generate";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faPlus } from "@fortawesome/pro-solid-svg-icons";
 import { TeamLink, UserLink } from "../../../../Links/Links";
