+++ packages/cyberstorm/src/components/Layout/Teams/TeamSettings/TeamSettingsLayout.tsx
@@ -7,8 +7,8 @@ import { Tabs } from "../../../Tabs/Tabs";
 import { TeamMembers } from "./TeamMembers/TeamMembers";
 import { TeamServiceAccounts } from "./TeamServiceAccounts/TeamServiceAccounts";
 import { TeamProfile } from "./TeamProfile/TeamProfile";
-import { getTeamSettingsDummyData } from "../../../../dummyData";
-import { TeamSettings } from "../../../../schema";
+import { getTeamSettingsDummyData } from "@thunderstore/dapper/src/implementations/dummy/generate";
+import { TeamSettings } from "@thunderstore/dapper/src/schema";
 import { BaseLayout } from "../../BaseLayout/BaseLayout";
 import { PageHeader } from "../../BaseLayout/PageHeader/PageHeader";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
