+++ packages/cyberstorm/src/components/Layout/Teams/TeamsLayout.tsx
@@ -7,7 +7,7 @@ import { SettingItem } from "../../SettingItem/SettingItem";
 import { Button } from "../../Button/Button";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faPlus } from "@fortawesome/pro-solid-svg-icons";
-import { getTeamPreviewDummyData } from "../../../dummyData";
+import { getTeamPreviewDummyData } from "@thunderstore/dapper/src/implementations/dummy/generate";
 import { TeamList } from "./TeamList/TeamList";
 import { Dialog } from "../../Dialog/Dialog";
 import { TextInput } from "../../TextInput/TextInput";
