--- packages/cyberstorm/src/dummyData/index.ts
@@ -1,1 +1,0 @@
-export * from "./generate";
