+++ packages/cyberstorm/src/index.ts
@@ -47,6 +47,3 @@ export * from "./components/Layout/SimplePages/InternalServerError/InternalServe
 export * from "./components/Layout/SimplePages/NotFound/NotFoundLayout";
 export * from "./components/Layout/SimplePages/PrivacyPolicy/PrivacyPolicyLayout";
 export * from "./components/Layout/SimplePages/TermsOfService/TermsOfServiceLayout";
-
-export * from "./dummyData/index";
-export * from "./schema/index";
